Move protocol definitions to types.py
