+++ src/sentry/scm/actions.py
@@ -13,8 +13,20 @@ from sentry.scm.private.helpers import (
     map_repository_model_to_repository,
 )
 from sentry.scm.private.ipc import record_count_metric
-from sentry.scm.private.provider import (
+from sentry.scm.private.rate_limit import RateLimitProvider
+from sentry.scm.types import (
     ALL_PROTOCOLS,
+    SHA,
+    ActionResult,
+    ArchiveFormat,
+    ArchiveLink,
+    BranchName,
+    BuildConclusion,
+    BuildStatus,
+    CheckRun,
+    CheckRunOutput,
+    Comment,
+    Commit,
     CompareCommitsProtocol,
     CreateBranchProtocol,
     CreateCheckRunProtocol,
@@ -38,6 +50,7 @@ from sentry.scm.private.provider import (
     DeletePullRequestCommentProtocol,
     DeletePullRequestCommentReactionProtocol,
     DeletePullRequestReactionProtocol,
+    FileContent,
     GetArchiveLinkProtocol,
     GetBranchProtocol,
     GetCheckRunProtocol,
@@ -58,34 +71,15 @@ from sentry.scm.private.provider import (
     GetPullRequestReactionsProtocol,
     GetPullRequestsProtocol,
     GetTreeProtocol,
-    MinimizeCommentProtocol,
-    Provider,
-    RequestReviewProtocol,
-    UpdateBranchProtocol,
-    UpdateCheckRunProtocol,
-    UpdatePullRequestProtocol,
-)
-from sentry.scm.private.rate_limit import RateLimitProvider
-from sentry.scm.types import (
-    SHA,
-    ActionResult,
-    ArchiveFormat,
-    ArchiveLink,
-    BranchName,
-    BuildConclusion,
-    BuildStatus,
-    CheckRun,
-    CheckRunOutput,
-    Comment,
-    Commit,
-    FileContent,
     GitBlob,
     GitCommitObject,
     GitRef,
     GitTree,
     InputTreeEntry,
+    MinimizeCommentProtocol,
     PaginatedActionResult,
     PaginationParams,
+    Provider,
     PullRequest,
     PullRequestCommit,
     PullRequestFile,
@@ -96,12 +90,16 @@ from sentry.scm.types import (
     Repository,
     RepositoryId,
     RequestOptions,
+    RequestReviewProtocol,
     ResourceId,
     Review,
     ReviewComment,
     ReviewCommentInput,
     ReviewEvent,
     ReviewSide,
+    UpdateBranchProtocol,
+    UpdateCheckRunProtocol,
+    UpdatePullRequestProtocol,
 )
